backend/requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
bcrypt